moved UserIcon component
