+++ components/Header.tsx
@@ -1,7 +1,7 @@
 import Image from 'next/image'
 import Link from 'next/link'
 import { useAuth } from '~/contexts/AuthContext'
-import UserIcon from './UserIcon'
+import UserIcon from './users/UserIcon'
 import { Menu } from '@headlessui/react'
 import { useRouter } from 'next/router'
 import ButtonSmall from './button/ButtonSmall'
