+++ pages/roadmaps/[roadmapId].tsx
@@ -6,7 +6,7 @@ import BarMiddle from '~/components/parts/BarMiddle'
 import BarBottom from '~/components/parts/BarBottom'
 import RoadmapStatus from '~/components/parts/RoadmapStatus'
 import Tag from '~/components/parts/Tag'
-import UserIcon from '~/components/UserIcon'
+import UserIcon from '~/components/users/UserIcon'
 import Trash from '~/components/parts/Trash'
 import Button from '~/components/button/Button'
 import useAspidaSWR from '@aspida/swr'
